Fix data type mess up - there shouldn't be a need for conversion at all + fix a graphic name - when i "fixed" the "slider.png missing" bug, i actually didn't do it better ;)

+++ src/logic/productionsite.cc
@@ -315,7 +315,7 @@ void ProductionSite::calc_statistics()
 			(m_statistics_buffer, sizeof(m_statistics_buffer),
 			 "%d%%",    percOk);
 
-	m_last_stat_percent = static_cast<int8_t>(percOk);
+	m_last_stat_percent = percOk;
 
 	m_statistics_changed = false;
 }
